Make non /dev/* args a fatal error
This is important in context of the -n option. Since -n does not print a hint on
which disk is a particular temp for, skipping non /dev/* args would mess the
output. With this fix the number of lines in output is always the same as the
number of args.